# contracts/integration-contract/openapi.json
{
  "components": {
    "schemas": {
      "CreateRunRequest": {
        "description": "A request to start one workflow run.",
        "properties": {
          "input": {
            "additionalProperties": true,
            "title": "Input",
            "type": "object"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Run Id",
            "type": "string"
          },
          "runtime": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime"
          }
        },
        "required": [
          "run_id",
          "input"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "RunResponse": {
        "description": "One run as the runner reports it.\n\n`result`, `error`, and `artifact_storage` are omitted rather than null until they exist,\nwhich is why the route serializes with `exclude_none`.",
        "properties": {
          "artifact_storage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunnerArtifactStorage"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "title": "Result"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "workflow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow"
          }
        },
        "required": [
          "status"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunStatusQuery": {
        "description": "Which runs a status query wants status tokens for.",
        "properties": {
          "run_ids": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Run Ids",
            "type": "array"
          }
        },
        "required": [
          "run_ids"
        ],
        "title": "RunStatusQuery",
        "type": "object"
      },
      "RunStatusResponse": {
        "properties": {
          "missing": {
            "description": "Requested run IDs the runner holds no record for.",
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunStatusToken"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "missing"
        ],
        "title": "RunStatusResponse",
        "type": "object"
      },
      "RunStatusToken": {
        "description": "One run's status, without its record.\n\nA polling caller asks whether anything changed far more often than it needs results, so\nthis carries no result: the full record is fetched once per terminal run.",
        "properties": {
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "description": "The run's status as the workflow gateway reports it.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunStatusToken",
        "type": "object"
      },
      "RunnerArtifactRef": {
        "description": "A stored diagnostic bundle, as `store_run_artifacts` reports it.",
        "properties": {
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "kind": {
            "const": "diagnostic_bundle",
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "uri": {
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "uri",
          "media_type",
          "digest",
          "size_bytes"
        ],
        "title": "RunnerArtifactRef",
        "type": "object"
      },
      "RunnerArtifactStorage": {
        "description": "Where a terminal run's diagnostics went, or why they are not there.\n\nThe runner only attaches this once a run reaches a terminal state, and reports a failed\nupload as a status rather than an error, because storage is diagnostic metadata and not\nthe business result.",
        "properties": {
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunnerArtifactRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "enum": [
              "available",
              "unavailable",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "RunnerArtifactStorage",
        "type": "object"
      },
      "RunnerError": {
        "description": "The envelope every runner error carries.",
        "properties": {
          "category": {
            "enum": [
              "input",
              "workflow",
              "llm",
              "runtime",
              "internal"
            ],
            "title": "Category",
            "type": "string"
          },
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "category",
          "code",
          "message",
          "retryable"
        ],
        "title": "RunnerError",
        "type": "object"
      },
      "RunnerErrorResponse": {
        "description": "An error body.\n\n`run_id` and `workflow` are present when the runner already knows which run the request\nconcerned, so a caller can attribute a failure without echoing its own request back.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RunnerError"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "workflow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow"
          }
        },
        "required": [
          "error"
        ],
        "title": "RunnerErrorResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Bearer token configured through RUNNER_SERVICE_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "sec-review-agents runner service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Healthz"
      }
    },
    "/v1/runs/status": {
      "post": {
        "operationId": "get_run_statuses_v1_runs_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStatusQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Statuses"
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run"
      }
    },
    "/v1/workflows/{workflow}/runs": {
      "post": {
        "operationId": "create_run_v1_workflows__workflow__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow",
            "required": true,
            "schema": {
              "enum": [
                "issue-review",
                "pull-request-review",
                "repository-review"
              ],
              "title": "Workflow",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunnerErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Run"
      }
    }
  }
}